requirements.txt
streamlit
openai
groq
python-dotenv
SpeechRecognition
pydub
pyttsx3
fluvio
requests